Update Event Service

+++ event-service/src/main/java/com/gyp/eventservice/services/impl/SeasonServiceImpl.java
@@ -2,11 +2,13 @@ package com.gyp.eventservice.services.impl;
 
 import java.util.List;
 
+import com.gyp.common.dtos.pagination.PaginatedDto;
 import com.gyp.common.enums.event.SeasonStatus;
 import com.gyp.common.exceptions.ResourceDuplicateException;
 import com.gyp.common.exceptions.ResourceNotFoundException;
 import com.gyp.common.services.ValidationService;
 import com.gyp.common.utils.PropertyName;
+import com.gyp.common.utils.SecurityUtils;
 import com.gyp.common.validators.criteria.ValidationInfo;
 import com.gyp.common.validators.rulecheck.CheckFactory;
 import com.gyp.eventservice.dtos.season.SeasonRequestDto;
@@ -15,8 +17,12 @@ import com.gyp.eventservice.entities.SeasonEntity;
 import com.gyp.eventservice.mappers.SeasonMapper;
 import com.gyp.eventservice.repositories.SeasonRepository;
 import com.gyp.eventservice.services.SeasonService;
+import com.gyp.eventservice.services.criteria.SeasonSearchCriteria;
+import com.gyp.eventservice.services.specifications.SeasonSpecification;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 @Slf4j
@@ -35,6 +41,16 @@ public class SeasonServiceImpl implements SeasonService {
 	}
 
 	@Override
+	public List<SeasonResponseDto> getAllSeasons(SeasonSearchCriteria criteria, PaginatedDto pagination) {
+		Specification<SeasonEntity> seasonSpecification = SeasonSpecification.createSearchSeasonSpecification(criteria);
+		Page<SeasonEntity> entities = seasonRepository.findAll(seasonSpecification, pagination.toPageable());
+		if(!entities.isEmpty()) {
+			return seasonMapper.toResponseDtoList(entities.getContent());
+		}
+		return null;
+	}
+
+	@Override
 	public SeasonResponseDto getSeasonById(String seasonId) throws ResourceNotFoundException {
 		var seasonEntity = seasonRepository.findById(seasonId).orElseThrow(
 				() -> new ResourceNotFoundException(String.format("Season with ID %s not found", seasonId))
@@ -45,7 +61,10 @@ public class SeasonServiceImpl implements SeasonService {
 	@Override
 	public SeasonResponseDto createSeason(SeasonRequestDto seasonRequestDto) {
 		checkDuplicate(seasonRequestDto);
+		String organizationId = SecurityUtils.getCurrentOrganizationId();
 		var seasonEntity = seasonMapper.toEntity(seasonRequestDto);
+		seasonEntity.setStatus(SeasonStatus.ACTIVE);
+		seasonEntity.setOrganizationId(organizationId);
 		var savedSeasonEntity = seasonRepository.save(seasonEntity);
 		return seasonMapper.toResponseDto(savedSeasonEntity);
 	}
@@ -53,11 +72,15 @@ public class SeasonServiceImpl implements SeasonService {
 	@Override
 	public SeasonResponseDto updateSeason(String seasonId, SeasonRequestDto seasonRequestDto)
 			throws ResourceNotFoundException {
+		String organizationId = SecurityUtils.getCurrentOrganizationId();
 		var existingSeason = seasonRepository.findById(seasonId)
 				.orElseThrow(
 						() -> new ResourceNotFoundException(String.format("Season with ID %s not found", seasonId)));
-		checkDuplicate(seasonRequestDto);
+		if(existingSeason == null) {
+			throw new ResourceNotFoundException("Season not found");
+		}
 		seasonMapper.updateEntityFromDto(seasonRequestDto, existingSeason);
+		existingSeason.setOrganizationId(organizationId);
 		var updatedSeasonEntity = seasonRepository.save(existingSeason);
 		return seasonMapper.toResponseDto(updatedSeasonEntity);
 	}
@@ -67,8 +90,10 @@ public class SeasonServiceImpl implements SeasonService {
 		var existingSeason = seasonRepository.findById(seasonId)
 				.orElseThrow(
 						() -> new ResourceNotFoundException(String.format("Season with ID %s not found", seasonId)));
+		if(existingSeason == null) {
+			throw new ResourceNotFoundException("Season not found");
+		}
 		seasonRepository.delete(existingSeason);
-		log.info("Season with ID {} deleted successfully", seasonId);
 	}
 
 	@Override
